// WeatherApp/ClientApp/src/app/Models/ExtractedWeatherProperties.ts
export class ExtractedWeatherProperties {
    hours: string[] = [];
    temperature: number[] = [];
    feelsLikeTemperature: number[] = [];
    windSpeed: number[] = [];
    windDirection: number[] = [];
    cloudinessLevel: number[] = [];
    humidity: number[] = [];
    pressure: number[] = [];
}